# std.format: New version for formatting integral values using writeAligned