I first tried to refactor the old implementation, but soon noticed, that this is a long path leading to nothing. So removed most of it and wrote it anew. What I kept is the handling of "%r" and the calculation of `base`, as well as some other smaller ideas.

What I skipped is: 
- Using three functions instead of one. Seems to me, that this was a relict from ancient times, not necessary anymore. 
- Support for `cent` and `ucent`. The old implementation would not have worked anyway and I see not need in supporting something, that might never be added. And if needed, it should not be a problem to make this work for `cent` too.
- Extra handling for one digit numbers. I checked this. With modern optimization techniques, this had no effect at all.

What changed:
- `%u` has been handled like `%d` and was not documented. I recently added it to the docs and handled it like `%b`, `%o` and `%x`. That means, the value is now cast to `unsigend` and treated like an unsigned, that is, not sign is printed, even if plus or space flag are present. That is consistent with the other three format characters and thus `%u` got a meaning. (I had to change a unittest though.)
- Formatting 0 with 0 precision resulted in an empty string. Two unittests checked for that, which suggests, that someone found that natural. I can't go with that. There is a big difference if I write "There are bugs" or "There are 0 bugs". I even considered this a bug, see issue 21820. So I changed these two unittests.

Speed (in ns, with ldc2 and optimization):
type | old | new
-|-|-
byte | 129 | 134
ubyte | 126 | 131
short | 161 | 157
ushort | 155  | 146
int | 235 | 233
uint | 234 | 204
long | 396 | 378
ulong | 393 | 342

I used a random sample of parameters for the format string and random numbers. Aside from bytes the new version is faster.

I also did these tests with `"%s"` and one digit values. The new version was about 1ns faster for all types (103ns vs. 102ns).