@berni44 2021-04-14
> Nice work, do you have any hypothesis on why bytes are slower?

Not really. They were a little bit faster in the 1 digit test but only about 0.3ns which could also be noise. If I should guess, I'd say, the optimizer did different optimizations. But it's out of the blue.